it is with great pleasure that I inform you we have ran the month
yet another u16 -> u32 fix here

kill event triggered
from client id #0 [$43051680/$1919568813/860527sh/1994228sh]
from client id #1 [$4271915616/$1915133226/718463sh/1987601sh]
mbo at 0x56421a18c
===raw DUMP===
===MBO DUMP===
MBO with 3091 levels and hi bid index 1481

make main  49.04s user 5.91s system 97% cpu 56.321 total

+++ src/ob.c
@@ -387,7 +387,7 @@ u32 ob_canrep(FL* orders, u32 order_id, void* old_mbo_raw, void* new_mbo_raw, CB
     u16 quantity = rep->quantity;
     u16 remaining_quantity = quantity;
     u8 direction = (rep->status >> BUY_DIRECTION_BIT) & 1;
-    u16 cancel_id = rep->other_id;
+    u32 cancel_id = rep->other_id;
     Order* can;
 
     printf("canrep %u\n", rep->status >> CAN_REP_BIT);
